% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlookHelpeR.R
\name{chainedTransfromations}
\alias{chainedTransfromations}
\title{Excute a chain of data transformations.}
\usage{
chainedTransfromations(data, ...)
}
\arguments{
\item{data}{A dataframe with data and a monthly period column}

\item{...}{Data transforming functions function(data){... returns(data)}}
}
\value{
A dataframe
}
\description{
\code{chainedTransfromations} returns the dataset with all transformation executed in order
}
