% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlookHelpeR.R
\name{laggedValue}
\alias{laggedValue}
\title{Generate a vector of a lagged value for a single column.}
\usage{
laggedValue(data, timeColumn, var, lag, period = "months")
}
\arguments{
\item{data}{A dataframe with data and a monthly period column}

\item{timeColumn}{The name of the period column (YYYY-MM-DD)}

\item{var}{The name of the variable to be lagged}

\item{lag}{The lag expressed in the number of months}

\item{period}{The period name (deafult = months)}
}
\value{
A vector
}
\description{
\code{laggedValue} returns a dataframe with a lagged column)
}
