% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlookHelpeR.R
\name{manyLaggedValues}
\alias{manyLaggedValues}
\title{Generate a series of lagged values.}
\usage{
manyLaggedValues(data, columnNames, timeColumn, lag, period = "months")
}
\arguments{
\item{data}{A dataframe with data and a monthly period column}

\item{columnNames}{Column names that are to be lagged}

\item{timeColumn}{The name of the period column (YYYY-MM-DD)}

\item{lag}{The lag expressed in the number of months}

\item{period}{The period name (deafult = months)}
}
\value{
A dataframe
}
\description{
\code{manyLaggedValues} returns the dataset with lagged values (suffix Lx where x is the lag number)
}
